détail event, validation event

+++ event_detail.php
@@ -13,9 +13,14 @@ include_once('layouts/header.php');?>
         <?php 
         if(!empty($_SESSION['id'])){
             //S'il n'est pas un participant, il a accès au bouton participer, sinon il peut annuler sa participation
-            if($est_participant===0){ ?>
-                <button onclick="location.href='php/ajouter_participation_evenement.php?id_event=<?php echo $_GET['id'];?>'">Participer</button>
+            if($est_participant===0){ 
+                //Vérification que le nombre de participants n'a pas atteint sa limite
+                if($nombre_participants!==$detail_evenement->limited_number_participant){ ?>
+                    <button onclick="location.href='php/ajouter_participation_evenement.php?id_event=<?php echo $_GET['id'];?>'">Participer</button>
             <?php }else{ ?>
+                        <div>Limite de participants atteinte !</div>
+                    <?php }
+            }else{ ?>
                 <button onclick="location.href='php/annuler_participation_evenement.php?id_event=<?php echo $_GET['id'];?>'">Annuler participation</button>
             <?php }
         }else{
@@ -26,10 +31,11 @@ include_once('layouts/header.php');?>
             <p>Rendez-vous à <?php echo $detail_evenement->name_city.', '.$detail_evenement->zip_code.', '.$detail_evenement->street_address.' de '.date('d-m-Y',strtotime($detail_evenement->date_start)).' à '.$detail_evenement->hour_start.' jusqu\'au '.date('d-m-Y',strtotime($detail_evenement->date_end)).' '.$detail_evenement->hour_end;?></p>
             <p><?php echo $detail_evenement->description;?></p>
         </div>
-        
-        <?php if($createur_evenement->id===$_SESSION['id'] || $_SESSION['role']==="Administrateur"){ ?>
-            <button onclick="location.href='php/supprimer_evenement.php?id_event=<?php echo $_GET['id'];?>'">Supprimer</button>
-        <?php } ?>
+        <?php if(!empty($_SESSION['id']) and !empty($_SESSION['role'])){ ?>
+            <?php if($createur_evenement->id===$_SESSION['id'] || $_SESSION['role']==="Administrateur"){ ?>
+                <button onclick="location.href='php/supprimer_evenement.php?id_event=<?php echo $_GET['id'];?>'">Supprimer</button>
+            <?php } 
+        }?>
     </div>
     <p>Nombre limitants de participants : <?php echo $detail_evenement->limited_number_participant;?></p>
     <?php if(empty($participants)){
@@ -47,7 +53,7 @@ include_once('layouts/header.php');?>
                     <tr>
                         <td><?php echo $participant->last_name;?></td>
                         <td><?php echo $participant->first_name;?></td>
-                        <td></td>
+                        <td><i onclick="location.href='profil.php?id=<?php echo $participant->id;?>'" class="fas fa-user"></i></td>
                     </tr>
                 <?php } ?>
             </tbody>
@@ -56,8 +62,9 @@ include_once('layouts/header.php');?>
     <h2>Espace commentaires</h2>
     <?php 
     if(!empty($_SESSION['id'])){ ?>
-        <form method="post">
-            <textarea placeholder="Votre commentaire ..."></textarea>
+        <form action="php/ajouter_commentaire.php" method="post">
+            <textarea name="commentaire" placeholder="Votre commentaire ..."></textarea>
+            <button name="bouton_event" value="<?php echo $_GET['id'];?>" type="submit">Envoyer</button>
         </form>
     <?php } 
     //Liste de tous les commentaires existants pour cet événement
@@ -71,12 +78,7 @@ include_once('layouts/header.php');?>
     
     <?php    
     /*
-        Si adminsitrateur ou créateur de l'event : bouton supprimer. (!empty $_session)
         Extension possible : Bouton modifier + popup confirmation
-        
-        Lister les participants
-        
-        Espace commentaires (si inscrit peut écrire, sinon lui indiquer qu'il doit se connecter, afficher les commentaires).
         Extension possible : suppression de son commentaire.
     */
